pyproject.toml
[project]
name = "experiments"
version = "0.0.0" # See current_version below
requires-python = "==3.13.*"
readme = "README.md"
dependencies = [
    "celery==5.5.2",
    "django==5.2",
    "django-celery-beat==2.8.0",
    "django-debug-toolbar==5.2.0",
    "faker==37.1.0",
    "gunicorn==23.0.0",
    "matplotlib==3.10.1",
    "psycopg[binary,pool]==3.2.7",
    "redis[hiredis]==6.0.0",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]==5.2.0",
    "fakeredis==2.28.1",
    "mypy==1.15.0",
    "pytest-django==4.11.1",
    "ruff==0.11.8",
]

[tool.uv]
package = false

# https://callowayproject.github.io/bump-my-version/
[tool.bumpversion]
allow_dirty = false
commit = true
message = "Release {new_version}"

# Update version and date in project source code
[[tool.bumpversion.files]]
filename = "config/__init__.py"
regex = true
search = """__version__ = \"[^"]*\"\n__date__ = \"[^"]*\""""
replace = """__version__ = \"{new_version}\"\n__date__ = \"{now:%Y-%m-%d}\""""

# Ruff: code formatter + linter
[tool.ruff]
line-length = 120

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings
docstring-code-format = true
docstring-code-line-length = 120

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "C", "Q", "B", "A", "PL", "I"]
ignore = []

[tool.ruff.lint.mccabe]
max-complexity = 10

# mypy
[tool.mypy]
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

# pytest
[tool.pytest.ini_options]
addopts = ["--ds=config.settings", "-vv"]
